- made descriptionWithLocale:indent: slightly more sensible - NSNull is now singleton - KVC now throws NSUndefinedKeyException - KVO now throws correct exceptions on unknown observed keys; cleans up in failure case - ObjCHashTable.h unix line endings; sensible string compare (before was downright wacky)

+++ Foundation/NSKeyValueCoding/NSKeyValueObserving-Private.h
@@ -14,7 +14,7 @@ THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLI
 	NSKeyValueObservingOptions options;
 	void* context;
 	id changeDictionary;
-	int willChangeCount;
+	NSUInteger willChangeCount;
 	NSString* keyPath;
 	id object;
 }
